Moved EventRepository to respository folder on back end. Changed routing on getAllEvents method
